Fix: - 컴포넌트와 비즈니스로직 분리 - 모달 객체로 분별
