+++ src/apis/updateCardData.ts
@@ -1,7 +1,7 @@
-import { PostCard } from 'src/types/cardTypes';
+import { PutCard } from 'src/types/cardTypes';
 import axiosInstance from './axiosInstance';
 
-export const updateCardData = async (cardId: number, body: PostCard) => {
+export const updateCardData = async (cardId: number, body: PutCard) => {
   const { data } = await axiosInstance.put(`cards/${cardId}`, body);
   return data;
 };
