+++ src/components/Modal/Card/CommentBox.tsx
@@ -1,10 +1,5 @@
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
-import { useForm } from 'react-hook-form';
-import deleteComment from 'src/apis/deleteComment';
-import postComment from 'src/apis/postComment';
-import readCommentList from 'src/apis/readCommentList';
 import { CardData } from 'src/types/cardTypes';
-import { CommentBody, PostComment } from 'src/types/commentTypes';
+import useCommentBox from 'src/hooks/useCommentBox';
 import CommentList from './CommentList';
 
 export default function CommentBox({
@@ -12,51 +7,8 @@ export default function CommentBox({
 }: {
   cardInformation: CardData;
 }) {
-  const queryClient = useQueryClient();
-  const { data: commentList } = useQuery<PostComment>({
-    queryKey: ['readCommentList', cardInformation.id],
-    queryFn: () => readCommentList(cardInformation.id)
-  });
-
-  const { mutate: createCommentMutate } = useMutation<void, Error, CommentBody>(
-    {
-      mutationFn: body => postComment(body),
-      onSuccess: () => {
-        queryClient.invalidateQueries({ queryKey: ['readCommentList'] });
-      }
-    }
-  );
-
-  const { mutate: deleteCommentMutate } = useMutation<
-    void,
-    Error,
-    { commentId: number }
-  >({
-    mutationFn: ({ commentId }) => deleteComment(commentId),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['readCommentList'] });
-    }
-  });
-
-  const deleteClick = (commentId: number) => {
-    deleteCommentMutate({ commentId });
-  };
-
-  const { register, handleSubmit, setValue } = useForm<{
-    content: string;
-    comment: number;
-  }>();
-
-  const submit = (data: { content: string }) => {
-    const body = {
-      ...data,
-      cardId: cardInformation.id,
-      columnId: cardInformation.columnId,
-      dashboardId: cardInformation.dashboardId
-    };
-    createCommentMutate(body);
-    setValue('content', '');
-  };
+  const { handleSubmit, submit, register, commentList, deleteClick } =
+    useCommentBox(cardInformation);
   return (
     <div>
       <form
